refactor: move sort_search_results to result_utils.py
